Move image chooser dialog fun
